Add events

+++ src/main/java/com/dreamteam/moneysplitter/service/EventService.java
@@ -4,11 +4,10 @@ import com.dreamteam.moneysplitter.assemblers.EventPurchaseResourceAssembler;
 import com.dreamteam.moneysplitter.assemblers.EventResourceAssembler;
 import com.dreamteam.moneysplitter.assemblers.OtherUserResourceAssembler;
 import com.dreamteam.moneysplitter.controller.EventController;
-import com.dreamteam.moneysplitter.domain.Event;
-import com.dreamteam.moneysplitter.domain.Purchase;
-import com.dreamteam.moneysplitter.domain.User;
-import com.dreamteam.moneysplitter.domain.UserEvent;
+import com.dreamteam.moneysplitter.controller.UserProfileController;
+import com.dreamteam.moneysplitter.domain.*;
 import com.dreamteam.moneysplitter.domain.dto.EventDTO;
+import com.dreamteam.moneysplitter.domain.dto.MyPartDTO;
 import com.dreamteam.moneysplitter.domain.dto.PurchaseDTO;
 import com.dreamteam.moneysplitter.domain.dto.UserDTO;
 import com.dreamteam.moneysplitter.repositories.EventRepo;
@@ -19,6 +18,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.hateoas.EntityModel;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.*;
 import java.util.stream.Collectors;
@@ -61,17 +61,17 @@ public class EventService {
         Map<String, Object> map = new HashMap<>();
         EventDTO eventDTO = DTOMaker.getEventDTO(event);
         Set<EntityModel<UserDTO>> collect = getUserDTOEntityModels(user, event);
-        Set<EntityModel<PurchaseDTO>> purchaseDTOS = GetPurchaseDTOEntityModels(user, event);
+        Set<EntityModel<PurchaseDTO>> purchaseDTOS = getPurchaseDTOEntityModels(user, event);
         map.put("event", eventResourceAssembler.toModel(eventDTO));
         map.put("users", collect);
         map.put("purchases", purchaseDTOS);
         return map;
     }
 
-    private Set<EntityModel<PurchaseDTO>> GetPurchaseDTOEntityModels(User user, Event event) {
+    private Set<EntityModel<PurchaseDTO>> getPurchaseDTOEntityModels(User user, Event event) {
         return purchaseRepo.findAllByEvent(event).stream().map(e -> {
             EntityModel<PurchaseDTO> purchaseDTOEntityModel = purchaseResourceAssembler.toModel(DTOMaker.getPurchaseDTO(e));
-            if(e.getUser().equals(user)){
+            if (e.getUser().equals(user)) {
                 purchaseDTOEntityModel.add(linkTo(methodOn(EventController.class).deletePurchaseFromEvent(event.getId(), e.getId()))
                         .withRel("DeletePurchase"));
             }
@@ -83,7 +83,7 @@ public class EventService {
         List<UserEvent> allByEvent = userEventRepo.findAllByEvent(event);
         return allByEvent.stream().map(e -> {
             EntityModel<UserDTO> userDTOEntityModel = friendResourceAssembler.toModel(DTOMaker.getUserDTO(e.getUser()));
-            if(event.getOwner().equals(user)){
+            if (event.getOwner().equals(user)) {
                 userDTOEntityModel.add(linkTo(methodOn(EventController.class).deleteUserFromEvent(e.getUser().getId(), event.getId()))
                         .withRel("DeletePerson"));
             }
@@ -106,7 +106,7 @@ public class EventService {
         event.setOwner(user);
         event.setIsClosed(false);
         Event save = eventRepo.save(event);
-        addUserToEvent(user, save);
+        addUserOwnToEvent(user, save);
         return eventResourceAssembler.toModel(DTOMaker.getEventDTO(save));
     }
 
@@ -114,8 +114,9 @@ public class EventService {
         User invitedUser = userRepo.findById(userId).orElseThrow();
         User inviter = userRepo.findByEmail(principal);
         Event event = eventRepo.findById(eventId).orElseThrow();
-        if (userEventRepo.findAllByEvent(event).stream().anyMatch(e -> e.getUser().equals(inviter))) {
-            addUserToEvent(invitedUser, event);
+        if (!event.getIsClosed() && isEventParticipant(inviter, event) &&
+                userEventRepo.findAllByEvent(event).stream().noneMatch(e -> e.getUser().equals(invitedUser))) {
+            addUserOwnToEvent(invitedUser, event);
             return true;
         }
         return false;
@@ -125,7 +126,7 @@ public class EventService {
         User deletedUser = userRepo.findById(userId).orElseThrow();
         User owner = userRepo.findByEmail(principal);
         Event event = eventRepo.findById(eventId).orElseThrow();
-        if (event.getOwner().equals(owner) && userEventRepo.findAllByEvent(event).stream().anyMatch(e -> e.getUser().equals(deletedUser))) {
+        if (!event.getIsClosed() && event.getOwner().equals(owner) && isEventParticipant(deletedUser, event)) {
             Optional<UserEvent> first = userEventRepo.findAllByEvent(event).stream().filter(e -> e.getUser().equals(deletedUser)).findFirst();
             if (first.isPresent()) {
                 userEventRepo.delete(first.get());
@@ -135,34 +136,69 @@ public class EventService {
         return false;
     }
 
-    private void addUserToEvent(User user, Event event) {
+    private void addUserOwnToEvent(User user, Event event) {
         UserEvent userEvent = new UserEvent();
         userEvent.setUser(user);
         userEvent.setEvent(event);
+        userEvent.setGeneralContribution(BigDecimal.ZERO);
+        userEvent.setId(new UserEventKey(user.getId(), event.getId()));
         userEventRepo.save(userEvent);
     }
 
     public boolean addPurchaseToEvent(Long eventId, PurchaseDTO purchaseDTO, String principal) {
         User user = userRepo.findByEmail(principal);
         Event event = eventRepo.findById(eventId).orElseThrow();
-        if(userEventRepo.findAllByEvent(event).stream().anyMatch(e -> e.getUser().equals(user))){
+        if (!event.getIsClosed() && isEventParticipant(user, event)) {
             Purchase purchase = new Purchase();
             purchase.setDate(LocalDate.now().toString());
             purchase.setEvent(event);
             purchase.setUser(user);
             purchase.setPurchaseCost(purchaseDTO.getPurchaseCost());
             purchase.setPurchaseName(purchaseDTO.getPurchaseName());
+            purchaseRepo.save(purchase);
             return true;
         }
         return false;
+    }
+
+    private boolean isEventParticipant(User user, Event event) {
+        return userEventRepo.findAllByEvent(event).stream().anyMatch(e -> e.getUser().equals(user));
     }
 
     public boolean deletePurchaseFromEvent(Long eventId, Long purchaseId, String principal) {
         User user = userRepo.findByEmail(principal);
         Event event = eventRepo.findById(eventId).orElseThrow();
         Purchase purchase = purchaseRepo.findById(purchaseId).orElseThrow();
-        if(purchase.getEvent() != null && purchase.getEvent().equals(event) && purchase.getUser().equals(user)){
+        if (!event.getIsClosed() && purchase.getEvent() != null && purchase.getEvent().equals(event) && purchase.getUser().equals(user)) {
             purchaseRepo.delete(purchase);
+            return true;
+        }
+        return false;
+    }
+
+    public EntityModel<MyPartDTO> getMyPart(String principal, Long eventId) {
+        Event event = eventRepo.findById(eventId).orElse(null);
+        User user = userRepo.findByEmail(principal);
+        if(event != null && !event.getIsClosed() && isEventParticipant(user, event)){
+            List<Purchase> allPurchasesByEvent = purchaseRepo.findAllByEvent(event);
+            BigDecimal allCost = allPurchasesByEvent.stream().map(Purchase::getPurchaseCost).reduce(BigDecimal.ZERO, BigDecimal::add);
+            BigDecimal myPart = allCost.divide(BigDecimal.valueOf(userEventRepo.findAllByEvent(event).size()));
+            return EntityModel.of(new MyPartDTO(myPart),
+                    linkTo(methodOn(UserProfileController.class).myProfile()).withRel("myAccount"),
+                    linkTo(methodOn(EventController.class).getMyEvents()).withRel("myEvents"),
+                    linkTo(methodOn(EventController.class).getEventById(eventId)).withRel("thisEvent"),
+                    linkTo(methodOn(EventController.class).split(eventId)).withSelfRel());
+        }
+        return null;
+    }
+
+    public boolean closeEvent(String principal, Long eventId) {
+        Event event = eventRepo.findById(eventId).orElse(null);
+        User user = userRepo.findByEmail(principal);
+        if(event != null && event.getOwner().equals(user)){
+            event.setIsClosed(true);
+            event.setDateClosed(LocalDate.now().toString());
+            eventRepo.save(event);
             return true;
         }
         return false;
